Translate timeline by mouse

+++ composeApp/src/desktopMain/kotlin/com/alekso/dltstudio/ui/MainWindow.kt
@@ -20,13 +20,13 @@ fun MainWindow() {
     var dltSession by remember { mutableStateOf<ParseSession?>(null) }
     var progress by remember { mutableStateOf(0f) }
     var tabIndex by remember { mutableStateOf(0) }
-    var offset by remember { mutableStateOf(0) }
+    var offset by remember { mutableStateOf(0f) }
     var scale by remember { mutableStateOf(1f) }
 
     val tabClickListener: (Int) -> Unit = { i -> tabIndex = i }
     val statusBarProgressCallback: (Float) -> Unit = { i -> progress = i }
     val newSessionCallback: (ParseSession) -> Unit = { newSession -> dltSession = newSession }
-    val offsetUpdateCallback: (Int) -> Unit = { newOffset -> offset = newOffset }
+    val offsetUpdateCallback: (Float) -> Unit = { newOffset -> offset = newOffset }
     val scaleUpdateCallback: (Float) -> Unit = { newScale -> scale = newScale }
 
     Column {
